+++ include/valueset.h
@@ -1,6 +1,6 @@
 /*
  *	valueset.h -- interface to value->stuff mapping routines
- *	Copyright (C) 2008-2013 Fred Barnes <frmb@kent.ac.uk>
+ *	Copyright (C) 2008-2016 Fred Barnes <frmb@kent.ac.uk>
  *
  *	This program is free software; you can redistribute it and/or modify
  *	it under the terms of the GNU General Public License as published by
@@ -31,10 +31,10 @@ typedef enum ENUM_setstrategy {
 } setstrategy_e;
 
 typedef struct TAG_valueset {
-	DYNARRAY (int, values);
+	DYNARRAY (int64_t, values);
 	DYNARRAY (struct TAG_tnode *, links);
-	int v_min, v_max;
-	int v_base, v_limit;
+	int64_t v_min, v_max;
+	int64_t v_base, v_limit;
 	setstrategy_e strat;
 } valueset_t;
 
@@ -46,7 +46,7 @@ extern valueset_t *valueset_create (void);
 extern void valueset_free (valueset_t *vset);
 extern void valueset_dumptree (valueset_t *vset, int indent, struct TAG_fhandle *stream);
 
-extern int valueset_insert (valueset_t *vset, int val, struct TAG_tnode *link);
+extern int valueset_insert (valueset_t *vset, int64_t val, struct TAG_tnode *link);
 extern int valueset_decide (valueset_t *vset);
 extern int valueset_sort (valueset_t *vset);
 extern int valueset_insertblanks (valueset_t *vset, struct TAG_tnode *link);
